+++ Apple.h
@@ -4,15 +4,15 @@
 #include <SFML/Graphics.hpp>
 
 class Apple {
-    float AppleX_;
-    float AppleY_;
+    int AppleX_;
+    int AppleY_;
     std::mt19937 gen_;
 public:
     Apple();
     void randomizePosition();
     [[nodiscard]] sf::RectangleShape drawApple() const;
-    [[nodiscard]] float AppleX() const { return AppleX_; }
-    [[nodiscard]] float AppleY() const { return AppleY_; }
+    [[nodiscard]] int AppleX() const { return AppleX_; }
+    [[nodiscard]] int AppleY() const { return AppleY_; }
 
 };
 #endif //SNAKESFML_APPLE_H
